+++ src/main/java/ru/softwerke/practice/app2019/storage/Unique.java
@@ -1,7 +1,5 @@
 package ru.softwerke.practice.app2019.storage;
 
-import java.util.UUID;
-
 public interface Unique {
-    int getId();
+    long getId();
 }
